feat(rust): move base64 implementation to own file, update cargo
